+++ src/components/base/b-router/CHANGELOG.md
@@ -13,7 +13,7 @@ Changelog
 
 #### :rocket: New Feature
 
-* Added new event `linkNavigate` to `b-router` to provide the capability to prevent router transition upon link click
+* Added new event `hrefTransition` to `b-router` to provide the capability to prevent router transition upon link click
 
 ## v4.0.0-beta.7 (2023-06-27)